Updating all the components preview to tsx
