+++ website/screens/theme-generator/components/Mode.tsx
@@ -3,34 +3,36 @@ import styled from "styled-components";
 
 type ModeProps = {
   text: string;
-  mode: string;
+  mode?: "light" | "dark";
   children: React.ReactNode;
 };
-const Mode = ({ text, mode, children }: ModeProps): JSX.Element => (
+const Mode = ({ text, mode = "light", children }: ModeProps): JSX.Element => (
   <ModeContainer mode={mode}>
     <Title mode={mode}>{text}</Title>
-    <PreviewsContainer>{children}</PreviewsContainer>
+    <PreviewContainer>{children}</PreviewContainer>
   </ModeContainer>
 );
 
 const ModeContainer = styled.div<{ mode: string }>`
+  display: flex;
+  flex-direction: column;
+  gap: 36px;
   background-color: ${(props) =>
     props.mode === "dark" ? "#000000" : "transparent"};
-  padding-bottom: 10px;
 `;
 
-const PreviewsContainer = styled.div`
-  display: flex;
-  justify-content: flex-start;
-  padding-left: 10px;
-`;
-
-const Title = styled.div<{ mode: string }>`
+const Title = styled.span<{ mode: string }>`
   font: Bold 12px/17px Open Sans;
+  line-height: 18px;
   letter-spacing: 1.88px;
   color: ${(props) => (props.mode === "dark" ? "#FFFFFF" : "#000000")};
   text-transform: uppercase;
-  padding: 20px 0 10px 10px;
+`;
+
+const PreviewContainer = styled.div`
+  display: flex;
+  gap: 5rem;
+  padding-left: 1rem;
 `;
 
 export default Mode;
