+++ website/screens/theme-generator/components/previews/Button.tsx
@@ -1,0 +1,54 @@
+import React from "react";
+import { DxcButton } from "@dxc-technology/halstack-react";
+import Mode from "../Mode";
+import linkedinIcon from "../../images/LinkedinIcon";
+import ExamplesContainer from "./ExamplesContainer";
+
+const Button = () => {
+  const onClick = () => {
+    console.log("click");
+  };
+
+  return (
+    <ExamplesContainer>
+      <Mode text="Primary">
+        <DxcButton mode="primary" label="button" />
+        <DxcButton
+          mode="primary"
+          label="Button"
+          icon={linkedinIcon}
+          onClick={onClick}
+        />
+        <DxcButton mode="primary" label="button" disabled onClick={onClick} />
+      </Mode>
+      <Mode text="Secondary">
+        <DxcButton mode="secondary" label="button" onClick={onClick} />
+        <DxcButton
+          label="Button"
+          mode="secondary"
+          icon={linkedinIcon}
+          iconPosition="after"
+          onClick={onClick}
+        />
+        <DxcButton
+          mode="secondary"
+          label="button"
+          disabled={true}
+          onClick={onClick}
+        />
+      </Mode>
+      <Mode text="Text">
+        <DxcButton mode="text" label="button" onClick={onClick} />
+        <DxcButton
+          label="Button"
+          mode="text"
+          icon={linkedinIcon}
+          onClick={onClick}
+        />
+        <DxcButton mode="text" label="button" disabled onClick={onClick} />
+      </Mode>
+    </ExamplesContainer>
+  );
+};
+
+export default Button;
