+++ website/screens/theme-generator/components/previews/Dialog.tsx
@@ -1,12 +1,26 @@
 import React, { useState } from "react";
-import styled from "styled-components";
 import {
   DxcDialog,
   DxcButton,
   DxcHeading,
   DxcTextInput,
+  DxcFlex,
+  DxcParagraph,
 } from "@dxc-technology/halstack-react";
 import Mode from "../Mode";
+import ExamplesContainer from "./ExamplesContainer";
+
+const errorIcon = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    height="24px"
+    viewBox="0 0 24 24"
+    width="24px"
+    fill="red"
+  >
+    <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-2h2v2zm0-4h-2V7h2v6z" />
+  </svg>
+);
 
 const Dialog = () => {
   const [isDefaultDialogVisible, setDefaultDialogVisible] = useState(false);
@@ -23,60 +37,29 @@ const Dialog = () => {
     setNoOverlayDialogVisible(!isNoOverlayDialogVisible);
   };
 
-  const ErrorIcon = () => (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      height="24px"
-      viewBox="0 0 24 24"
-      width="24px"
-      fill="red"
-    >
-      <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-2h2v2zm0-4h-2V7h2v6z" />
-    </svg>
-  );
-
   return (
-    <DialogContainer>
+    <ExamplesContainer>
       <Mode text="Default">
-        <DxcButton
-          label="Delete"
-          onClick={handleClickDefaultDialog}
-          margin={{ top: "xsmall" }}
-        ></DxcButton>
+        <DxcButton label="Delete" onClick={handleClickDefaultDialog} />
         {isDefaultDialogVisible && (
           <DxcDialog onCloseClick={handleClickDefaultDialog}>
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "row",
-                alignContent: "center",
-              }}
-            >
-              <span
-                style={{
-                  display: "flex",
-                  flexWrap: "wrap",
-                  alignContent: "center",
-                  marginRight: "10px",
-                }}
-              >
-                <ErrorIcon />
-              </span>
+            <DxcFlex alignContent="center" gap="1rem">
+              <DxcFlex wrap="wrap" alignContent="center">
+                {errorIcon}
+              </DxcFlex>
               <DxcHeading level={5} text="Delete document" />
-            </div>
-            <p>Are you sure you want to delete the 3 selected documents?</p>
-            <p style={{ textAlign: "right" }}>
+            </DxcFlex>
+            <DxcParagraph>
+              Are you sure you want to delete the 3 selected documents?
+            </DxcParagraph>
+            <DxcFlex alignItems="flex-end" gap="1rem">
               <DxcButton
                 label="Cancel"
                 onClick={handleClickDefaultDialog}
                 mode="text"
-                margin={{ top: "medium", right: "xxsmall" }}
-              ></DxcButton>
-              <DxcButton
-                label="Delete"
-                onClick={handleClickDefaultDialog}
-              ></DxcButton>
-            </p>
+              />
+              <DxcButton label="Delete" onClick={handleClickDefaultDialog} />
+            </DxcFlex>
           </DxcDialog>
         )}
       </Mode>
@@ -84,33 +67,24 @@ const Dialog = () => {
         <DxcButton
           label="See account information"
           onClick={handleClickBackground}
-          margin={{ top: "xsmall" }}
-        ></DxcButton>
+        />
         {isModalDialogVisible && (
           <DxcDialog padding="medium" onBackgroundClick={handleClickBackground}>
-            <DxcHeading level={4} text="Account information" />
-            <DxcHeading level={5} text="Client" margin={{ top: "xsmall" }} />
-            <p>
+            <DxcFlex direction="column" gap="1rem">
+              <DxcHeading level={4} text="Account information" />
+              <DxcHeading level={5} text="Client" />
               <DxcTextInput label="Name" />
-            </p>
-            <p>
               <DxcTextInput label="Surname" />
-            </p>
-            <p>
               <DxcTextInput label="Address" />
-            </p>
-            <p style={{ textAlign: "right" }}>
+            </DxcFlex>
+            <DxcFlex alignItems="flex-end">
               <DxcButton
                 label="Cancel"
                 onClick={handleClickBackground}
                 mode="text"
-                margin={{ top: "medium", right: "xxsmall" }}
-              ></DxcButton>
-              <DxcButton
-                label="Add client"
-                onClick={handleClickBackground}
-              ></DxcButton>
-            </p>
+              />
+              <DxcButton label="Add client" onClick={handleClickBackground} />
+            </DxcFlex>
           </DxcDialog>
         )}
       </Mode>
@@ -127,14 +101,12 @@ const Dialog = () => {
             onCloseClick={handleClickNoOverlay}
             overlay={false}
           >
-            <p>An example of Dialog without darker background (overlay).</p>
+            <DxcParagraph>An example of Dialog without darker background (overlay).</DxcParagraph>
           </DxcDialog>
         )}
       </Mode>
-    </DialogContainer>
+    </ExamplesContainer>
   );
 };
-
-const DialogContainer = styled.div``;
 
 export default Dialog;
