+++ website/screens/theme-generator/components/previews/ProgressBar.tsx
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import styled from "styled-components";
 import { DxcButton, DxcProgressBar } from "@dxc-technology/halstack-react";
 import Mode from "../Mode";
+import ExamplesContainer from "./ExamplesContainer";
 
 const ProgressBar = () => {
   const [isVisible, changeIsVisible] = useState(false);
@@ -12,7 +13,7 @@ const ProgressBar = () => {
     });
   };
   const fetchData = () => {
-    return new Promise((resolve) => {
+    return new Promise<void>((resolve) => {
       setTimeout(() => {
         resolve();
       }, 3000);
@@ -20,14 +21,13 @@ const ProgressBar = () => {
   };
 
   return (
-    <ProgressBarContainer>
+    <ExamplesContainer>
       <Mode text="Undeterminate default">
         <Container>
           <DxcProgressBar
             label="Loading"
             helperText="Helper text"
             overlay={false}
-            margin={{ top: "xsmall", bottom: "xxsmall" }}
           />
         </Container>
       </Mode>
@@ -39,17 +39,14 @@ const ProgressBar = () => {
             overlay={false}
             showValue
             value={45}
-            margin={{ top: "xsmall", bottom: "xxsmall" }}
           />
         </Container>
       </Mode>
       <Mode text="With overlay">
         <DxcButton
-          theme="light"
           label="Show Progress Bar for 3 seconds"
           onClick={showModal}
         />
-
         {isVisible && (
           <DxcProgressBar
             label="Loading"
@@ -58,37 +55,9 @@ const ProgressBar = () => {
           />
         )}
       </Mode>
-
-      {/* <h4>Dark mode</h4>
-      <BackgroundColorProvider color="#000000">
-        <Mode mode="dark" text="Undeterminate default">
-          <Container>
-            <DxcProgressBar
-              label="Loading"
-              helperText="Helper text"
-              overlay={false}
-              margin={{ top: "xsmall", bottom: "xxsmall" }}
-            />
-          </Container>
-        </Mode>
-        <Mode mode="dark" text="Determinate default">
-          <Container>
-            <DxcProgressBar
-              label="Loading"
-              helperText="Helper text"
-              overlay={false}
-              showValue
-              value={45}
-              margin={{ top: "xsmall", bottom: "xxsmall" }}
-            />
-          </Container>
-        </Mode>
-      </BackgroundColorProvider> */}
-    </ProgressBarContainer>
+    </ExamplesContainer>
   );
 };
-
-const ProgressBarContainer = styled.div``;
 
 const Container = styled.div`
   width: 100%;
