--- website/screens/theme-generator/components/previews/Switch.jsx
@@ -1,74 +1,0 @@
-import React, { useState } from "react";
-import styled from "styled-components";
-import { DxcSwitch } from "@dxc-technology/halstack-react";
-import Mode from "../Mode";
-
-const Switch = () => {
-  const [checked, changeChecked] = useState(false);
-
-  const onChange = (newValue) => {
-    changeChecked(newValue);
-  };
-
-  return (
-    <SwitchContainer>
-      <Mode text="Default">
-        <DxcSwitch
-          checked={checked}
-          label="Switch"
-          onChange={onChange}
-          margin={{ top: "xsmall", left: "xsmall" }}
-        />
-      </Mode>
-      <Mode text="Disabled">
-        <DxcSwitch
-          checked={true}
-          label="Label before"
-          margin={{ top: "xsmall", left: "xsmall", right: "xsmall" }}
-          disabled
-        />
-        <DxcSwitch
-          checked={false}
-          labelPosition="after"
-          label="Label after"
-          margin={{ top: "xsmall", left: "xsmall" }}
-          disabled
-        />
-      </Mode>
-      {/* <DxcHeading
-        text="Dark Mode"
-        level={5}
-        margin={{ top:"xsmall", bottom: "xxsmall" }}
-      />
-      <BackgroundColorProvider color="#000000">
-        <Mode mode="dark" text="Default">
-          <DxcSwitch
-            checked={checkedDark}
-            label="Switch"
-            onChange={onChangeDark}
-            margin={{ top: "xsmall", left: "xsmall" }}
-          />
-        </Mode>
-        <Mode mode="dark" text="Disabled">
-          <DxcSwitch
-            checked={true}
-            label="Label before"
-            margin={{ top: "xsmall", left: "xsmall", right: "xsmall" }}
-            disabled
-          />
-          <DxcSwitch
-            checked={false}
-            labelPosition="after"
-            label="Label after"
-            margin={{ top: "xsmall", left: "xsmall" }}
-            disabled
-          />
-        </Mode>
-      </BackgroundColorProvider> */}
-    </SwitchContainer>
-  );
-};
-
-const SwitchContainer = styled.div``;
-
-export default Switch;
